Show calendar
Добавлен фрагмент календаря

+++ ToDoAppANDROID/app/src/main/java/com/example/todoappandroid/EditTaskFragment.kt
@@ -168,7 +168,7 @@ class EditTaskFragment : Fragment() {
 
                         // ← ШАГ 3: Создаём обновлённую задачу
                         val updatedTask = Task(
-                            id = taskId,
+                            id = taskId.toInt(),
                             title = title,
                             description = if (description.isNotEmpty()) description else null,
                             date = if (date.isNotEmpty()) date else null,
